Extract Group widget and add unit tests
